requirements.txt
numpy
streamlit
pandas
pillow
mistralai
pictex